Refactor how ripple animation is applied to `GestureHandlerButton` (#2014)
This PR changes how the ripple animation is applied to `RNGestureHandlerButton` on Android (see the linked issue to see the problem). I don't really know what was causing the issue but it's possible that it was a problem with underlying Android implementation - the existing code worked correctly on Android 12 (API 32).

The new implementation creates `RippleDrawable` programmatically instead of using `selectableItemBackground` and `selectableItemBackgroundBorderless` attributes.

`applyRippleEffectWhenNeeded` was removed, as it was used to set the ripple radius and color to the drawable, but now it's done during its creation in updated `createSelectableDrawable`.

Fixes https://github.com/software-mansion/react-native-gesture-handler/issues/618.