Fix state flow for manual activation on iOS (#1912)
On iOS when using `manualActivation`, the gesture would receive `ACTIVATE` event when the state wasn't changed. It happened because the `UIGestureRecognizer` used to prevent it from activating would fail, fulfilling the requirement of `shouldBeRequiredToFailByGestureRecognizer`. This PR changes the behavior of the blocker to activate instead of failing, thus causing the blocked gesture to fail if it wasn't activated manually.

Fixes https://github.com/software-mansion/react-native-gesture-handler/issues/1865.