Fix importing `PressabilityDebugView` on web (#2109)
`PressabilityDebugView` is not implemented in `react-native-web` so importing it in `createHandler.tsx` was causing errors on web. This PR moves the import to a separate file with a web-specific version that returns `null`, thus preventing errors.

I also added the `key` property when rendering the `PressabilityDebugView` to prevent errors about it.