Make gestures work on non-editable `TextInput` on Android (#2112)
Closes https://github.com/software-mansion/react-native-gesture-handler/issues/774.

This PR makes gestures work on `TextInputs` with `editable` property set to `false` by extracting the handlers attached to the `EditText` even when it's disabled (by setting `editable=false`). On native side, its `EditText` has `isEnabled` property set to `false`. Because of that it fell into `PointerEventsConfig.BOX_NONE` branch of `when` expression and handlers attached to it weren't extracted (mainly the `NativeViewGestureHandler`). This caused the RN's gesture system to take over and call `requestDisallowInterceptTouchEvent` causing other gestures to fail.