Pass `disabled` prop to the BaseButton in TouchableNativeFeedback (#2141)
The `disabled` prop on Touchables from Gesture Handler was not passed to the `BaseButton`, instead it was used to ignore incoming events. This successfully made them not call `onPress*()` callbacks, but in case of `TouchableNativeFeedback` it didn't disable the event handling on the native side.
This PR makes `GenericTouchable` pass `disabled` prop further down to `BaseButton`, which in turn passes it to the native view, disabling touch interaction on it.