Detect duplicate instances (#2155)
When a project exists with more than one instance of Gesture Handler, two versions of Gesture Handler JS module try to use the same Native Module. It can occur when some of your dependencies have installed Gesture Handler inside their own node_modules instead of using it as a peer dependency. The problem is, for example, when the old version of JS tries to call a removed function in a newer version of the native module or can call twice initialization of the native module and override existing data. The solution to this issue is to modify the package.json file, requiring only one Gesture Handler version.

Co-authored-by: Jakub Piasecki <jakub.piasecki@swmansion.com>