Export custom `RefreshControl` component and add logic required to handle it (#2137)
Closes https://github.com/software-mansion/react-native-gesture-handler/issues/1067.

The problem with `RefreshControl` on Android was that it wasn't a part of the Gesture Handler system. This PR adds a custom `RefreshControl` component exported by GH in the same way as `ScrollView`, `Switch` etc. It also adds custom logic to `ScrollView` and `FlatList` components exported by Gesture Handler that handles setting relations between native gestures properly.