Check if loaded module is really Reanimated in `reanimatedWrapper` (#2179)
In some cases `reanimatedWrapper` would not fail when loading `react-native-reanimated` even though it was not installed, instead loading some other file/module. In my case it was `utils` file. This caused exception when calling `useAnimatedGesture` in `GestureDetector` as methods required by it were not present in the wrapper.

This PR updates `reanimatedWrapper` so that it checks for existence of `useSharedValue` on the loaded module and setting it back to `undefined` if it's not present.