Manual gesture receiving touch events (#2188)
## Description

This PR introduces first step to make `ManualGestureHandler` available on web. Touch events are now being sent to handlers with prop `needsPointerData`, so it's possible to interact with them. However, there's still no option to manually manage state of this handler, as it requires more changes in state management logic.

## Test plan

Tested on example app