Transform entire `MotionEvent` instead of single pointer only (#2156)
## Description

Gesture Handler on Android intercepts events via the
`GestureHandlerRootView`, which means that we need to transform the
touch coordinated by ourselves to check if it's inside the view.
Similarly, we need to transform the event when dispatching it to one of
the gesture handlers in order to correctly calculate the changes and
send touch events.

The current method works well until there is more than one pointer at
once on the screen. When there is more than one pointer, only the first
one will be correctly transformed, while others will simply be moved by
the offset calculated when transforming the first pointer.

This PR changes the logic when delivering events to gesture handlers -
the whole event is transformed, and handlers receive both events -
transformed and untransformed one. This way things like touch events and
`NativeViewGestureHandler` which rely on the location in the coordinate
space of the view keep working, and other gestures get to work in the
coordinate space of the root view.

Fixes
https://github.com/software-mansion/react-native-gesture-handler/issues/2138.

## Test plan

Tested on the Example app.