Fix `NullPointerException` when evaluating `hitSlop` on Android (#2241)
## Description

Fixes
https://github.com/software-mansion/react-native-gesture-handler/issues/2239.

Based on the location of the crash from the issue this seems to be a
concurrency issue. Using `?.let` instead of non-null assertions should
resolve the problem, as the used variable will be guaranteed not to
change while the block is being executed.

## Test plan

I don't really know how to reproduce it, but I also don't see what could
be the other solution, so... trust me?