Make handlers respect `shouldCancelWhenOutside` (#2271)
## Description

This PR solves web part of #2265 issue. I've modified `onPointerOut` method to act similar as on android. Right now, if pointer leaves handler and `shouldCancelWhenOutside` is set to `true`, handler goes to `cancel` state if it was `active`, or to `fail`, if it was in `began`

This change also affects `ManualGestureHandler`, which now cancels properly.

## Test plan

Tested on example app