Merge `lib` package into the main one in the Android implementation (#2270)
## Description

This PR changes the structure of the Gesture Handler's Android code. For
some time GH could have been used as a native Android library for
handling gestures, but it has been removed since. The split into two
packages that was required by it is still there, along some of the now
redundant files (another `build.gradle` and `AndroidManifest`).

This PR removes the two-package split and moves the core files into a
new `core` subpackage, like the `react` one.

## Test plan

Build the example app.