Fix missing `begin` event on Android for Pinch and Rotation (#2264)
## Description

This PR sets the focal point/anchor to the position of the pointer
before sending `begin` event on Android. At the position it's being
sent, it's always `NaN` at the moment, which causes the event not to be
received by the detector.

## Test plan

Add pinch & rotation gestures to the app and check whether the `begin`
event is called.