Reformat iOS codebase (#2296)
## Description

This PR reformats the iOS codebase to be consistent.

The intention was to drag this in time - when someone modifies an iOS
file, it would be formatted so the history would be less impacted. But
it turns out that it also formats files while merging branches, which is
pretty irritating. Turns out you can straight up [ignore certain
commits](https://docs.github.com/en/repositories/working-with-files/using-files/viewing-a-file#ignore-commits-in-the-blame-view)
when using `git blame` - this solves the problem IMO.

`.git-blame-ignore-revs` will need to be updated with the correct commit
hash after merging.