Show error when gestures are used without root view (#2420)
## Description

Android requires using GestureHandlerRootView to intercept events,
without it the gestures don't work and buttons hang in the "in-between"
state of pressed and not pressed.

This PR adds logic responsible for checking if gestures are attached to
views mounted under GestureHandlerRootView. If they aren't, the error
message is shown.

Supersedes
https://github.com/software-mansion/react-native-gesture-handler/pull/2115.

## Test plan

Tested on the Example & FabricExample app.