Fix Gesture Handlers getting stuck due to missing `up` event when using the new web implementation (#2442)
## Description

Sometimes the view didn't receive the `pointerup` event when moving the
pointer really fast, even though it was captured. This was causing a
problem where the gesture would never be finished, preventing every
other handler from activating.

Funny thing is, despite not receiving the `pointerup` event,
`lostpointercapture` event was still there. I used this fact to fill the
missing `up` event, so the gesture can correctly handle it.

I've only been able to reproduce it in Chrome, when using external
display.

## Test plan

Tested on the Example app, specifically the `Draggable` example.

---------

Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>