Fix Gesture Handler's buttons activating after scroll when using `RefreshControl` (#2457)
## Description

Closes
https://github.com/software-mansion/react-native-gesture-handler/issues/2351

`RefreshControl` from GH is set up so that `ScrollView` wrapped with it
should wait for it to fail. This way the scroll handler does not cancel
the `RefreshControl`. The problem with this approach is that the
`RefreshControl` handler always stays active preventing scroll from
activating.

This PR adds a workaround for this problem using
`SwipeRefreshLayoutHook`, which will cancel the `RefreshControl` if the
`ScrollView` is not at the top during the swipe.

## Test plan

Tested on the snack from
https://github.com/software-mansion/react-native-gesture-handler/issues/2351