Fix operation scheduling on iOS (#2483)
## Description


https://github.com/software-mansion/react-native-gesture-handler/pull/2467
replaces calls to `setImmediate` and `requestAnimationFrame` with
`queueMicrotask`. This introduces a bug, where on iOS the
`flushOperations` method would get called too early.

This PR changes `flushOperations` to only be used on the new
architecture, where it's actually required, and adds a retry mechanism
for attaching handlers to native views. It also updates parts of the new
arch-related code to function properly - finding the recognizer
responsible for the touch responder and attaching the root view handler.

Fixes
https://github.com/software-mansion/react-native-gesture-handler/issues/2482

## Test plan

Test on the Example and FabricExample apps.

---------

Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>