Fix root view error for jest (#2491)
## Description
This PR fixed a problem related to `HandlerName must be used as a
descendant of GestureHandlerRootView. Otherwise the gestures will not be
recognized. See
https://docs.swmansion.com/react-native-gesture-handler/docs/installation
for more details.` in jest.
<!--
Description and motivation for this PR.
In some project 

Include 'Fixes #<number>' if this is fixing some issue.
-->

## Related issues


[2488](https://github.com/software-mansion/react-native-gesture-handler/issues/2488)

## Test plan

tested locally on my project :)