Prevent GH buttons from activating when scrolling list (#2494)
## Description

If Gesture Handler buttons were inside a list, they could activate
during the scroll, which is undesirable. This PR makes them cancel when
the pointer moves more than the touch slop, preventing the activation.

It also adds a sanity check to the `onPointerOut` callback, so the
buttons won't be canceled twice, in between calls to `reset`, which was
breaking the state logic and causing them to be unresponsive every
second time.

## Test plan

Tested on the example app.