`activeCursor` prop (#2550)
## Description

This PR adds `activeCursor` prop which allows users to change cursor
appearance upon handler activation. Initially it was added in
[2.6.1](https://github.com/software-mansion/react-native-gesture-handler/releases/tag/2.6.1)
as a response to [this
issue](https://github.com/software-mansion/react-native-gesture-handler/issues/700).
However, not everyone liked the idea (which is understandable), so we
decided to turn this change into property. This way anyone who was using
it still will be able to do so, and those who didn't like the change
don't have to remove it on their own.

Co-authored-by: Michał Bert <michal.bert@swmansion.com>