Move REACT_NATIVE_VERSION to native-only code (#2666)
## Description

Fixes #2660

Depending on this package in a web-only app is causing a compile error.
This is due to this package importing `react-native/package.json`, which
for web-only builds doesn't exist. This import is only done to obtain
the react-native version number, which isn't used for web anyway. So it
seems easy enough to move this version number to a native-only file.

I chose to export a **function** so that it's a _logic error_ to use
`getReactNativeVersion` on web, and an error is thrown (right now it's
only used in one place, behind a check that makes sure it runs on native
only). It might have been easier to just export a **const**, with fake
values on web, but that would hide the error for a potentially more
confusing error down the line.

## Test plan

Not tested, sorry. Would appreciate if someone could help out here.

---------

Co-authored-by: Chris Coomber <ccoomber@whiterosemaths.com>