Fix inputs focus on safari (#2675)
## Description

It seems that since release of new web implementations `select` doesn't work as it should on Safari. The problem is caused by `setPointerCapture` method. Even though it is called on all browsers, input stays focused only on Safari. 

This PR changes method of calling `setPointerCapture` - now we do this only for `Div` elements.

Fixes #2673

## Test plan

Tested on example app