Add a new gesture relation - `blocksHandlers` - working like `reversed waitFor` (#2627)
## Description

Introduces a new relation between gestures, which allows to set one
gesture recognizer as blocking for another one. It's basically the same
as `waitFor` but in the other direction - `waitFor` allows to specify
which gestures must fail before this one can activate. The new prop
allows to specify which gestures should require this one to fail before
they can activate.

For the name, I think we've settled at `blocksHandlers` for the old API
and `.blocksExternalGesture()` for the new API.

There's no need to add another type of composition since it accomplishes
the same thing as `Gesture.Exclusive`, and is inherently a
between-components relation.

## Test plan

Tested on the Example app

Since it relies on already existing logic, I assume it will work
correctly. I would like to do more testing nonetheless.
Also needs to be tested on physical iOS device with multi-touch
gestures.

---------

Co-authored-by: Kacper Kapuściak <39658211+kacperkapusciak@users.noreply.github.com>