Support React Native 0.73 (#2671)
## Description

- updates `CMakeLists.txt` so includes can be resolved
- moves architecture-specific files out of `src` - at first, I thought
that it would resolve the redefinition errors, but it seems like a clean
build is required after changing the architecture. Anyway, since it's
already there and unifies how source sets are handled I figured it may
as well stay.

Updates root directory and `FabricExample` to RN 0.73. I've run into an
issue with babel trying to upgrade the `example` app (some `export` were
not transpiled to `module.exports`)

## Test plan

Tested on FabricExample app (both architectures).