Change functions in windows module file to no-ops. (#2741)
## Description

A while ago we added `RNGestureHandlerModule.windows.ts` file, to allow usage of our library on Windows. The problem is that it was an exact copy of `.web.ts` file, therefore `HTMLElement` is not defined. This PR changes all functions inside this module file into no-ops, so that applications will build correctly. 


## Test plan

Confirmed that app builds correctly on Windows