Stop sending event on pressure changes (#2762)
## Description

When the pressure changes, the stylus sends a `pointermove` event. Because it is highly sensitive to these changes, these events were being sent constantly. This PR introduces a check to see whether the pointer has changed its position. If it hasn't, the `pointermove` events won't be sent.

## Test plan

Tested on example app (draggable example and `console.log` inside `onPointerMove` inside `Pan`)