Add `mouseButton` implementation on Android (#2680)
## Description

This PR adds implementation of `mouseButton` prop on android.

Since `actionButton` field is available only in API >= 23, we provide full support only for those versions. The only thing that changes for API < 23 is ignoring `ACTION_BUTTON_*`, because we don't want handlers to react to both type of events.

Note that it requires [this PR](https://github.com/software-mansion/react-native-gesture-handler/pull/2676) to work.

## Test plan

Tested on `MouseButtons` example.