Add `TypeDoc` to new API (#2874)
## Description

I thought that adding TypeDocs might be a good idea and it could enhance
developer experience (I've come up with this idea when I discovered that
we have `withRef` modifier). This PR adds TypeDocs to new API (but we
can also think of filling missing docs in old API)

## Test plan

Hover functions/components and read docs.