Get right bridge (#2886)
## Description

This PR fixes getting access to right bridge, because `currentBridge`
method can return invalid pointer when more than one instance of react
exists in the application. Same changes we made in the Reanimated -
https://github.com/software-mansion/react-native-reanimated/pull/5953

## Test plan

I tested FabricExample on iOS with/without bridge, after startup and
after reload, and it seems to work.