docs: add @swmansion/t-rex-ui (#2895)
## Description

Because of a ramping up maintainance cost across our suite of tools -
Reanimated, Gesture Handler & React Native IDE we've decided to create a
UI library for our reusable docusaurus components.

The
[`@swmansion/t-rex-ui`](https://github.com/software-mansion-labs/t-rex-ui)
package aims to unify the components across our documentations. This
will make the maintainance easier in the long term by code reuse. We
hope that with a single source of truth will avoid the need of fixing a
bug in 3-4 places instead of just one.

Sidenote: the package requires the React & ReactDOM in version `18` or
higher.

## Test plan

```
cd docs/
yarn

# to test development build
yarn start

# to test production build
yarn build
yarn serve
```

---------

Co-authored-by: Jakub Piasecki <jakub.piasecki@swmansion.com>